Split DOM utils and translations into their own modules
