+++ _js/dom.js
@@ -1,0 +1,37 @@
+export function elem() {
+  const name = arguments[0]
+  const attrs = arguments.length === 3 ? arguments[1] : {}
+  const content = arguments.length === 3 ? arguments[2] : arguments[1]
+  const installBlock = document.createElement(name)
+  Object.keys(attrs).forEach(key => {
+    installBlock.setAttribute(key, attrs[key])
+  })
+  append(installBlock, content)
+  return installBlock
+}
+
+export function append(parent, content) {
+  if (content === undefined || content === null) {
+    return
+  }
+  switch (typeof content) {
+    case 'string':
+      parent.appendChild(document.createTextNode(content))
+      break
+    case 'object':
+      if (Array.isArray(content)) {
+        content.forEach(c => {
+          append(parent, c)
+        })
+        break
+      }
+    default:
+      parent.appendChild(content)
+  }
+}
+
+export function clear(myNode) {
+  while (myNode.firstChild) {
+    myNode.removeChild(myNode.firstChild)
+  }
+}
