+++ js/plugins.js
@@ -70,7 +70,11 @@ var t,e=arguments[3],r=require("process");!function(e,r){"object"==typeof export
 "use strict";module.exports=require("./").polyfill();
 },{"./":"Zt7E"}],"MCp7":[function(require,module,exports) {
 "use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.Headers=a,exports.Request=m,exports.Response=A,exports.fetch=x,exports.DOMException=void 0;var t={searchParams:"URLSearchParams"in self,iterable:"Symbol"in self&&"iterator"in Symbol,blob:"FileReader"in self&&"Blob"in self&&function(){try{return new Blob,!0}catch(t){return!1}}(),formData:"FormData"in self,arrayBuffer:"ArrayBuffer"in self};function e(t){return t&&DataView.prototype.isPrototypeOf(t)}if(t.arrayBuffer)var r=["[object Int8Array]","[object Uint8Array]","[object Uint8ClampedArray]","[object Int16Array]","[object Uint16Array]","[object Int32Array]","[object Uint32Array]","[object Float32Array]","[object Float64Array]"],o=ArrayBuffer.isView||function(t){return t&&r.indexOf(Object.prototype.toString.call(t))>-1};function n(t){if("string"!=typeof t&&(t=String(t)),/[^a-z0-9\-#$%&'*+.^_`|~]/i.test(t))throw new TypeError("Invalid character in header field name");return t.toLowerCase()}function s(t){return"string"!=typeof t&&(t=String(t)),t}function i(e){var r={next:function(){var t=e.shift();return{done:void 0===t,value:t}}};return t.iterable&&(r[Symbol.iterator]=function(){return r}),r}function a(t){this.map={},t instanceof a?t.forEach(function(t,e){this.append(e,t)},this):Array.isArray(t)?t.forEach(function(t){this.append(t[0],t[1])},this):t&&Object.getOwnPropertyNames(t).forEach(function(e){this.append(e,t[e])},this)}function h(t){if(t.bodyUsed)return Promise.reject(new TypeError("Already read"));t.bodyUsed=!0}function u(t){return new Promise(function(e,r){t.onload=function(){e(t.result)},t.onerror=function(){r(t.error)}})}function f(t){var e=new FileReader,r=u(e);return e.readAsArrayBuffer(t),r}function d(t){var e=new FileReader,r=u(e);return e.readAsText(t),r}function l(t){for(var e=new Uint8Array(t),r=new Array(e.length),o=0;o<e.length;o++)r[o]=String.fromCharCode(e[o]);return r.join("")}function c(t){if(t.slice)return t.slice(0);var e=new Uint8Array(t.byteLength);return e.set(new Uint8Array(t)),e.buffer}function y(){return this.bodyUsed=!1,this._initBody=function(r){this._bodyInit=r,r?"string"==typeof r?this._bodyText=r:t.blob&&Blob.prototype.isPrototypeOf(r)?this._bodyBlob=r:t.formData&&FormData.prototype.isPrototypeOf(r)?this._bodyFormData=r:t.searchParams&&URLSearchParams.prototype.isPrototypeOf(r)?this._bodyText=r.toString():t.arrayBuffer&&t.blob&&e(r)?(this._bodyArrayBuffer=c(r.buffer),this._bodyInit=new Blob([this._bodyArrayBuffer])):t.arrayBuffer&&(ArrayBuffer.prototype.isPrototypeOf(r)||o(r))?this._bodyArrayBuffer=c(r):this._bodyText=r=Object.prototype.toString.call(r):this._bodyText="",this.headers.get("content-type")||("string"==typeof r?this.headers.set("content-type","text/plain;charset=UTF-8"):this._bodyBlob&&this._bodyBlob.type?this.headers.set("content-type",this._bodyBlob.type):t.searchParams&&URLSearchParams.prototype.isPrototypeOf(r)&&this.headers.set("content-type","application/x-www-form-urlencoded;charset=UTF-8"))},t.blob&&(this.blob=function(){var t=h(this);if(t)return t;if(this._bodyBlob)return Promise.resolve(this._bodyBlob);if(this._bodyArrayBuffer)return Promise.resolve(new Blob([this._bodyArrayBuffer]));if(this._bodyFormData)throw new Error("could not read FormData body as blob");return Promise.resolve(new Blob([this._bodyText]))},this.arrayBuffer=function(){return this._bodyArrayBuffer?h(this)||Promise.resolve(this._bodyArrayBuffer):this.blob().then(f)}),this.text=function(){var t=h(this);if(t)return t;if(this._bodyBlob)return d(this._bodyBlob);if(this._bodyArrayBuffer)return Promise.resolve(l(this._bodyArrayBuffer));if(this._bodyFormData)throw new Error("could not read FormData body as text");return Promise.resolve(this._bodyText)},t.formData&&(this.formData=function(){return this.text().then(w)}),this.json=function(){return this.text().then(JSON.parse)},this}a.prototype.append=function(t,e){t=n(t),e=s(e);var r=this.map[t];this.map[t]=r?r+", "+e:e},a.prototype.delete=function(t){delete this.map[n(t)]},a.prototype.get=function(t){return t=n(t),this.has(t)?this.map[t]:null},a.prototype.has=function(t){return this.map.hasOwnProperty(n(t))},a.prototype.set=function(t,e){this.map[n(t)]=s(e)},a.prototype.forEach=function(t,e){for(var r in this.map)this.map.hasOwnProperty(r)&&t.call(e,this.map[r],r,this)},a.prototype.keys=function(){var t=[];return this.forEach(function(e,r){t.push(r)}),i(t)},a.prototype.values=function(){var t=[];return this.forEach(function(e){t.push(e)}),i(t)},a.prototype.entries=function(){var t=[];return this.forEach(function(e,r){t.push([r,e])}),i(t)},t.iterable&&(a.prototype[Symbol.iterator]=a.prototype.entries);var p=["DELETE","GET","HEAD","OPTIONS","POST","PUT"];function b(t){var e=t.toUpperCase();return p.indexOf(e)>-1?e:t}function m(t,e){var r=(e=e||{}).body;if(t instanceof m){if(t.bodyUsed)throw new TypeError("Already read");this.url=t.url,this.credentials=t.credentials,e.headers||(this.headers=new a(t.headers)),this.method=t.method,this.mode=t.mode,this.signal=t.signal,r||null==t._bodyInit||(r=t._bodyInit,t.bodyUsed=!0)}else this.url=String(t);if(this.credentials=e.credentials||this.credentials||"same-origin",!e.headers&&this.headers||(this.headers=new a(e.headers)),this.method=b(e.method||this.method||"GET"),this.mode=e.mode||this.mode||null,this.signal=e.signal||this.signal,this.referrer=null,("GET"===this.method||"HEAD"===this.method)&&r)throw new TypeError("Body not allowed for GET or HEAD requests");this._initBody(r)}function w(t){var e=new FormData;return t.trim().split("&").forEach(function(t){if(t){var r=t.split("="),o=r.shift().replace(/\+/g," "),n=r.join("=").replace(/\+/g," ");e.append(decodeURIComponent(o),decodeURIComponent(n))}}),e}function v(t){var e=new a;return t.replace(/\r?\n[\t ]+/g," ").split(/\r?\n/).forEach(function(t){var r=t.split(":"),o=r.shift().trim();if(o){var n=r.join(":").trim();e.append(o,n)}}),e}function A(t,e){e||(e={}),this.type="default",this.status=void 0===e.status?200:e.status,this.ok=this.status>=200&&this.status<300,this.statusText="statusText"in e?e.statusText:"OK",this.headers=new a(e.headers),this.url=e.url||"",this._initBody(t)}m.prototype.clone=function(){return new m(this,{body:this._bodyInit})},y.call(m.prototype),y.call(A.prototype),A.prototype.clone=function(){return new A(this._bodyInit,{status:this.status,statusText:this.statusText,headers:new a(this.headers),url:this.url})},A.error=function(){var t=new A(null,{status:0,statusText:""});return t.type="error",t};var _=[301,302,303,307,308];A.redirect=function(t,e){if(-1===_.indexOf(e))throw new RangeError("Invalid status code");return new A(null,{status:e,headers:{location:t}})};var E=self.DOMException;exports.DOMException=E;try{new E}catch(g){exports.DOMException=E=function(t,e){this.message=t,this.name=e;var r=Error(t);this.stack=r.stack},E.prototype=Object.create(Error.prototype),E.prototype.constructor=E}function x(e,r){return new Promise(function(o,n){var s=new m(e,r);if(s.signal&&s.signal.aborted)return n(new E("Aborted","AbortError"));var i=new XMLHttpRequest;function a(){i.abort()}i.onload=function(){var t={status:i.status,statusText:i.statusText,headers:v(i.getAllResponseHeaders()||"")};t.url="responseURL"in i?i.responseURL:t.headers.get("X-Request-URL");var e="response"in i?i.response:i.responseText;o(new A(e,t))},i.onerror=function(){n(new TypeError("Network request failed"))},i.ontimeout=function(){n(new TypeError("Network request failed"))},i.onabort=function(){n(new E("Aborted","AbortError"))},i.open(s.method,s.url,!0),"include"===s.credentials?i.withCredentials=!0:"omit"===s.credentials&&(i.withCredentials=!1),"responseType"in i&&t.blob&&(i.responseType="blob"),s.headers.forEach(function(t,e){i.setRequestHeader(e,t)}),s.signal&&(s.signal.addEventListener("abort",a),i.onreadystatechange=function(){4===i.readyState&&s.signal.removeEventListener("abort",a)}),i.send(void 0===s._bodyInit?null:s._bodyInit)})}x.polyfill=!0,self.fetch||(self.fetch=x,self.Headers=a,self.Request=m,self.Response=A);
+},{}],"fRxd":[function(require,module,exports) {
+"use strict";function e(t){return(e="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(t)}function t(){var e=arguments[0],t=3===arguments.length?arguments[1]:{},r=3===arguments.length?arguments[2]:arguments[1],n=document.createElement(e);return Object.keys(t).forEach(function(e){n.setAttribute(e,t[e])}),o(n,r),n}function o(t,r){if(null!=r)switch(e(r)){case"string":t.appendChild(document.createTextNode(r));break;case"object":if(Array.isArray(r)){r.forEach(function(e){o(t,e)});break}default:t.appendChild(r)}}function r(e){for(;e.firstChild;)e.removeChild(e.firstChild)}Object.defineProperty(exports,"__esModule",{value:!0}),exports.elem=t,exports.append=o,exports.clear=r;
+},{}],"1El3":[function(require,module,exports) {
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=n;var e={INSTALL_OLD:"DITA-OT 3.1 and older",INSTALL_CURRENT:"DITA-OT 3.2 and newer",LICENSE:"License",HOMEPAGE:"Homepage",KEYWORDS:"Keywords",INSTALL:"Install",FILTER_PLACEHOLDER:"Filter plugins",FILTER_ANY_VERSION:"Any version",FILTER_VERSION_LABEL:"DITA-OT version",NO_MATCHES:"No matches found.",DEPENDENCIES:"Dependencies",VERSIONS:"Versions",VERSION_NOT_FOUND:"Plugin {} version {} not found.",NOT_FOUND:"Plugin {} not found.",FOUND:"Found {} matches."};function n(n){return arguments.length>1?Array.prototype.slice.call(arguments).slice(1).reduce(function(e,n){return e.replace("{}",n)},e[n]):e[n]}
 },{}],"d5ui":[function(require,module,exports) {
-"use strict";var e=require("./version"),t=n(require("object.values"));function n(e){return e&&e.__esModule?e:{default:e}}function r(e){return(r="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function o(e,t){return c(e)||a(e,t)||i()}function i(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}function a(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(s){o=!0,i=s}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}function c(e){if(Array.isArray(e))return e}require("es6-promise/auto"),require("whatwg-fetch"),Object.values||t.default.shim();var s={INSTALL_OLD:"DITA-OT 3.1 and older",INSTALL_CURRENT:"DITA-OT 3.2 and newer",LICENSE:"License",HOMEPAGE:"Homepage",KEYWORDS:"Keywords",INSTALL:"Install",FILTER_PLACEHOLDER:"Filter plugins",FILTER_ANY_VERSION:"Any version",FILTER_VERSION_LABEL:"DITA-OT version",NO_MATCHES:"No matches found.",DEPENDENCIES:"Dependencies",VERSIONS:"Versions",VERSION_NOT_FOUND:"Plugin {} version {} not found.",NOT_FOUND:"Plugin {} not found.",FOUND:"Found {} matches."};function l(e){return arguments.length>1?Array.prototype.slice.call(arguments).slice(1).reduce(function(e,t){return e.replace("{}",t)},s[e]):s[e]}var u="https://plugins.dita-ot.org/_all.json",f=["3.2","3.1","3.0","2.5","2.4","2.3","2.2","2.1","2.0","1.8"],d=null;function p(e){d=e,Object.values(d).filter(function(e){return!!e}).forEach(function(e){e.alias||e.forEach(function(e){var t=e.name+" "+(e.description&&e.description)+" "+(e.keywords&&e.keywords.join(" "));e.search=t.toLocaleLowerCase().replace(/\W/g," ").replace(/\s+/g," ")})}),window.onpopstate=function(e){m(location.hash)},m(location.hash)}function m(t){var n=null;if(t){var r=o(t.substr(1).split("/"),2),i=r[0],a=r[1],c=d[i].slice();c.sort(e.compareVersion),n=T(c,a)||h(i,a)}else n=I(d);var s=document.getElementById("plugins");_(s),D(s,n),t||N()}function h(e,t){return w("p",t?l("VERSION_NOT_FOUND",e,t):l("NOT_FOUND",e))}document.addEventListener("DOMContentLoaded",function(e){fetch(u).then(function(e){return e.json()}).then(p).catch(function(e){console.error("Failed to fetch plugins: "+e)})});var y={freetext:null,version:null};function v(e){y.freetext=e.target.value.toLocaleLowerCase().replace(/\W/g," ").replace(/\s+/g," ").trim(),N()}function E(e){y.version=e.target.value,N()}function N(){if(y.freetext||y.version){var e=0;document.querySelectorAll("#list > li").forEach(function(t){L(d[t.id][0])?(e++,t.style.display="list-item"):t.style.display="none"});var t=document.querySelector("#empty"),n=document.querySelector("#hits");0===e?(t.style.display="block",n.style.display="none",_(n)):(t.style.display="none",n.style.display="block",_(n),n.appendChild(document.createTextNode(l("FOUND",e))))}else S()}function L(t){var n=!y.freetext;y.freetext&&t.search.match(y.freetext)&&(n=!0);var r=!y.version,o=t.deps.find(function(e){return"org.dita.base"===e.name});return y.version&&o&&(0,e.matchVersion)(y.version,o.req)&&(r=!0),n&&r}function S(){document.querySelectorAll("#list > li").forEach(function(e){e.style.display="list-item"}),document.querySelector("#hits").style.display="none"}function O(e){27===e.keyCode&&(document.querySelector("#query").value="",y.freetext="",N())}function g(){var e=w("input",{id:"query",value:y.freetext||"",type:"text",class:"form-control",placeholder:l("FILTER_PLACEHOLDER"),size:50},void 0);e.oninput=v,e.onkeypress=O;var t=f.map(function(e){var t={value:e};return e===y.version&&(t.selected="selected"),w("option",t,e)}),n=w("select",{id:"version",class:"form-control"},[w("option",{value:""},l("FILTER_ANY_VERSION"))].concat(t));return n.onchange=E,w("div",{class:"form-inline"},[w("div",{class:"form-group"},e)," ",w("div",{class:"form-group"},n)])}function I(e){return[g(),w("p",{id:"empty",style:"display: none; margin-top: 1em",class:"alert alert-info"},l("NO_MATCHES")),w("p",{id:"hits",style:"display: none; margin-top: 1em"},""),w("ul",{class:"list-unstyled",id:"list"},Object.values(e).filter(function(e){return!!e&&!e.alias}).sort(function(e,t){return e[0].name.localeCompare(t[0].name)}).map(function(e){return e[0]}).map(function(e){return w("li",{id:e.name},[w("h3",w("a",{href:"#".concat(e.name)},e.name)),w("p",e.description),w("p",(e.keywords||[]).reduce(function(e,t){return e.concat([w("code",{class:"small"},t),"  "])},[]))])}))]}function T(e,t){var n=t?e.find(function(e){return e.vers===t}):e[0];if(!n)return null;var r=document.createElement("div");r.appendChild(w("h2",["".concat(n.name),w("small"," ".concat(n.vers))])),n.description&&D(r,w("p",{class:"shortdesc"},n.description)),n.keywords&&0!==n.keywords.length&&D(r,[w("h3",l("KEYWORDS")),w("p",n.keywords.reduce(function(e,t){return e.concat([w("code",t),"  "])},[]))]),n.license&&D(r,[w("h3",l("LICENSE")),w("p",A(n.license))]),n.homepage&&D(r,[w("h3",l("HOMEPAGE")),w("p",w("a",{href:n.homepage},C(n.homepage)))]),D(r,[w("h3",l("INSTALL")),w("p",{class:"small"},l("INSTALL_CURRENT")),w("pre","dita --install ".concat(n.name)),w("p",{class:"small"},l("INSTALL_OLD")),w("pre","dita --install ".concat(n.url))]);var o=n.deps;return o.sort(function(e,t){return e.name.localeCompare(t.name)}),D(r,[w("h3",l("DEPENDENCIES")),w("ul",o.filter(function(e){return"org.dita.base"===e.name}).map(function(e){return w("li","DITA-OT ".concat(b(e.req)||""))})),w("ul",o.filter(function(e){return"org.dita.base"!==e.name}).map(function(e){return w("li","".concat(e.name," ").concat(b(e.req)||""))}))]),D(r,[w("h3",l("VERSIONS")),w("ul",e.map(function(e){return w("li",w("a",{href:"#".concat(n.name,"/").concat(e.vers)},e.vers))}))]),r}function A(e){switch(e){case"Apache-2.0":return w("a",{href:"https://www.apache.org/licenses/LICENSE-2.0"},"Apache License 2.0");case"MIT":return w("a",{href:"https://opensource.org/licenses/MIT"},"MIT License");case"UNLICENSED":return"Unlicensed";default:return e}}function b(e){return e.startsWith("=")?"".concat(e.substr(1)):e.startsWith(">=")?"".concat(e.substr(2)," or newer"):e.startsWith("<=")?"".concat(e.substr(2)," or older"):e}function C(e){return e.replace(/^\w+:\/\/([^\/]+?)(\/.*)?$/,"$1")}function w(){var e=arguments[0],t=3===arguments.length?arguments[1]:{},n=3===arguments.length?arguments[2]:arguments[1],r=document.createElement(e);return Object.keys(t).forEach(function(e){r.setAttribute(e,t[e])}),D(r,n),r}function D(e,t){if(null!=t)switch(r(t)){case"string":e.appendChild(document.createTextNode(t));break;case"object":if(Array.isArray(t)){t.forEach(function(t){D(e,t)});break}default:e.appendChild(t)}}function _(e){for(;e.firstChild;)e.removeChild(e.firstChild)}
-},{"./version":"RvaL","object.values":"Jlaz","es6-promise/auto":"KHzD","whatwg-fetch":"MCp7"}]},{},["d5ui"], null)
+"use strict";var e=require("./version"),t=l(require("object.values"));require("es6-promise/auto"),require("whatwg-fetch");var n=require("./dom"),r=l(require("./translations"));function l(e){return e&&e.__esModule?e:{default:e}}function a(e,t){return i(e)||c(e,t)||o()}function o(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}function c(e,t){var n=[],r=!0,l=!1,a=void 0;try{for(var o,c=e[Symbol.iterator]();!(r=(o=c.next()).done)&&(n.push(o.value),!t||n.length!==t);r=!0);}catch(i){l=!0,a=i}finally{try{r||null==c.return||c.return()}finally{if(l)throw a}}return n}function i(e){if(Array.isArray(e))return e}Object.values||t.default.shim();var s="https://plugins.dita-ot.org/_all.json",u=["3.2","3.1","3.0","2.5","2.4","2.3","2.2","2.1","2.0","1.8"],d=null;function m(e){d=e,Object.values(d).filter(function(e){return!!e}).forEach(function(e){e.alias||e.forEach(function(e){var t=e.name+" "+(e.description&&e.description)+" "+(e.keywords&&e.keywords.join(" "));e.search=t.toLocaleLowerCase().replace(/\W/g," ").replace(/\s+/g," ")})}),window.onpopstate=function(e){f(location.hash)},f(location.hash)}function f(t){var r=null;if(t){var l=a(t.substr(1).split("/"),2),o=l[0],c=l[1],i=d[o].slice();i.sort(e.compareVersion),r=I(i,c)||p(o,c)}else r=S(d);var s=document.getElementById("plugins");(0,n.clear)(s),(0,n.append)(s,r),t||E()}function p(e,t){return(0,n.elem)("p",t?(0,r.default)("VERSION_NOT_FOUND",e,t):(0,r.default)("NOT_FOUND",e))}document.addEventListener("DOMContentLoaded",function(e){fetch(s).then(function(e){return e.json()}).then(m).catch(function(e){console.error("Failed to fetch plugins: "+e)})});var h={freetext:null,version:null};function v(e){h.freetext=e.target.value.toLocaleLowerCase().replace(/\W/g," ").replace(/\s+/g," ").trim(),E()}function y(e){h.version=e.target.value,E()}function E(){if(h.freetext||h.version){var e=0;document.querySelectorAll("#list > li").forEach(function(t){g(d[t.id][0])?(e++,t.style.display="list-item"):t.style.display="none"});var t=document.querySelector("#empty"),l=document.querySelector("#hits");0===e?(t.style.display="block",l.style.display="none",(0,n.clear)(l)):(t.style.display="none",l.style.display="block",(0,n.clear)(l),l.appendChild(document.createTextNode((0,r.default)("FOUND",e))))}else L()}function g(t){var n=!h.freetext;h.freetext&&t.search.match(h.freetext)&&(n=!0);var r=!h.version,l=t.deps.find(function(e){return"org.dita.base"===e.name});return h.version&&l&&(0,e.matchVersion)(h.version,l.req)&&(r=!0),n&&r}function L(){document.querySelectorAll("#list > li").forEach(function(e){e.style.display="list-item"}),document.querySelector("#hits").style.display="none"}function w(e){27===e.keyCode&&(document.querySelector("#query").value="",h.freetext="",E())}function N(){var e=(0,n.elem)("input",{id:"query",value:h.freetext||"",type:"text",class:"form-control",placeholder:(0,r.default)("FILTER_PLACEHOLDER"),size:50},void 0);e.oninput=v,e.onkeypress=w;var t=u.map(function(e){var t={value:e};return e===h.version&&(t.selected="selected"),(0,n.elem)("option",t,e)}),l=(0,n.elem)("select",{id:"version",class:"form-control"},[(0,n.elem)("option",{value:""},(0,r.default)("FILTER_ANY_VERSION"))].concat(t));return l.onchange=y,(0,n.elem)("div",{class:"form-inline"},[(0,n.elem)("div",{class:"form-group"},e)," ",(0,n.elem)("div",{class:"form-group"},l)])}function S(e){return[N(),(0,n.elem)("p",{id:"empty",style:"display: none; margin-top: 1em",class:"alert alert-info"},(0,r.default)("NO_MATCHES")),(0,n.elem)("p",{id:"hits",style:"display: none; margin-top: 1em"},""),(0,n.elem)("ul",{class:"list-unstyled",id:"list"},Object.values(e).filter(function(e){return!!e&&!e.alias}).sort(function(e,t){return e[0].name.localeCompare(t[0].name)}).map(function(e){return e[0]}).map(function(e){return(0,n.elem)("li",{id:e.name},[(0,n.elem)("h3",(0,n.elem)("a",{href:"#".concat(e.name)},e.name)),(0,n.elem)("p",e.description),(0,n.elem)("p",(e.keywords||[]).reduce(function(e,t){return e.concat([(0,n.elem)("code",{class:"small"},t),"  "])},[]))])}))]}function I(e,t){var l=t?e.find(function(e){return e.vers===t}):e[0];if(!l)return null;var a=document.createElement("div");a.appendChild((0,n.elem)("h2",["".concat(l.name),(0,n.elem)("small"," ".concat(l.vers))])),l.description&&(0,n.append)(a,(0,n.elem)("p",{class:"shortdesc"},l.description)),l.keywords&&0!==l.keywords.length&&(0,n.append)(a,[(0,n.elem)("h3",(0,r.default)("KEYWORDS")),(0,n.elem)("p",l.keywords.reduce(function(e,t){return e.concat([(0,n.elem)("code",t),"  "])},[]))]),l.license&&(0,n.append)(a,[(0,n.elem)("h3",(0,r.default)("LICENSE")),(0,n.elem)("p",O(l.license))]),l.homepage&&(0,n.append)(a,[(0,n.elem)("h3",(0,r.default)("HOMEPAGE")),(0,n.elem)("p",(0,n.elem)("a",{href:l.homepage},T(l.homepage)))]),(0,n.append)(a,[(0,n.elem)("h3",(0,r.default)("INSTALL")),(0,n.elem)("p",{class:"small"},(0,r.default)("INSTALL_CURRENT")),(0,n.elem)("pre","dita --install ".concat(l.name)),(0,n.elem)("p",{class:"small"},(0,r.default)("INSTALL_OLD")),(0,n.elem)("pre","dita --install ".concat(l.url))]);var o=l.deps;return o.sort(function(e,t){return e.name.localeCompare(t.name)}),(0,n.append)(a,[(0,n.elem)("h3",(0,r.default)("DEPENDENCIES")),(0,n.elem)("ul",o.filter(function(e){return"org.dita.base"===e.name}).map(function(e){return(0,n.elem)("li","DITA-OT ".concat(q(e.req)||""))})),(0,n.elem)("ul",o.filter(function(e){return"org.dita.base"!==e.name}).map(function(e){return(0,n.elem)("li","".concat(e.name," ").concat(q(e.req)||""))}))]),(0,n.append)(a,[(0,n.elem)("h3",(0,r.default)("VERSIONS")),(0,n.elem)("ul",e.map(function(e){return(0,n.elem)("li",(0,n.elem)("a",{href:"#".concat(l.name,"/").concat(e.vers)},e.vers))}))]),a}function O(e){switch(e){case"Apache-2.0":return(0,n.elem)("a",{href:"https://www.apache.org/licenses/LICENSE-2.0"},"Apache License 2.0");case"MIT":return(0,n.elem)("a",{href:"https://opensource.org/licenses/MIT"},"MIT License");case"UNLICENSED":return"Unlicensed";default:return e}}function q(e){return e.startsWith("=")?"".concat(e.substr(1)):e.startsWith(">=")?"".concat(e.substr(2)," or newer"):e.startsWith("<=")?"".concat(e.substr(2)," or older"):e}function T(e){return e.replace(/^\w+:\/\/([^\/]+?)(\/.*)?$/,"$1")}
+},{"./version":"RvaL","object.values":"Jlaz","es6-promise/auto":"KHzD","whatwg-fetch":"MCp7","./dom":"fRxd","./translations":"1El3"}]},{},["d5ui"], null)
 //# sourceMappingURL=/plugins.map
